refactor: migrate Timeline
